+++ acme/services/SemanticManager.py
@@ -45,6 +45,9 @@ class SemanticManager(object):
 		"""	Check that the *descriptor* attribute conforms to the syntax defined by
 			the *descriptorRepresentation* attribute. 
 
+			Todo:
+				Not fully implemented yet.
+
 			Args:
 				smd: SMD object to use in the validation.
 			Return:
@@ -60,18 +63,37 @@ class SemanticManager(object):
 		# TODO implement real validation
 		L.isWarn and L.logWarn('Validation of SMD.descriptor is not implemented')
 
+		# TODO in case of an erro: generate a Response Status Code indicating an "INVALID_SPARQL_QUERY" error
 		return Result.successResult()
 
 	
+	def validateSPARQL(self, query:str) -> Result:
+		"""	Validate wether an input string is a valid SPARQL query.
+
+			Todo:
+				Not implemented yet.
+
+			Args:
+				query: String with the SPARQL query to validate.
+			Return:
+				Result object indicating success or error.
+		"""
+
+		L.isWarn and L.logWarn('Validation of SMD.semanticOpExec is not implemented')
+		return Result.successResult()
+
 
 	def validateValidationEnable(self, smd:SMD) -> Result:
 		"""	Check and handle the setting of the *validationEnable* attribute.
+
+			Todo:
+				Not fully implemented yet.
 
 			Args:
 				smd: SMD object to use in the validation. **Attn**: This procedure might update and change the provided *smd* object.
 			Return:
 				Result object indicating success or error.
-		 """
+		"""
 		# TODO implement validation and setting of validationEnable attribute
 		# Procedure from TS-0004
 		# The Hosting CSE shall set the validationEnable attribute of the <semanticDescriptor> resource based on
@@ -87,5 +109,55 @@ class SemanticManager(object):
 		# - validationEnable attribute is set to true or false according to the value provided in the request.
 		L.isWarn and L.logWarn('Validation of SMD.validationEnable is not implemented')
 		smd.setAttribute('vlde', False)
+		smd.setAttribute('svd', False)
 
+		return Result.successResult()
+
+
+	def validateSemantic(self, smd:SMD) -> Result:
+		"""	Perform the semantic validation of the <SMD> resource
+
+			Todo:
+				Not fully implemented yet.
+
+			Args:
+				smd: SMD object to use in the validation. **Attn**: This procedure might update and change the provided *smd* object.
+			Return:
+				Result object indicating success or error.
+		"""
+
+		if smd.vlde:	# Validation enabled
+			...
+			# b) If the validationEnable attribute is set as true, the hosting CSE shall perform the semantic validation process in
+			# 	the following steps according to clause 7.10.2 in oneM2M TS-0034 [50]. Otherwise, skip the following steps.
+			# c) Check if the addressed <semanticDescriptor> resource is linked to other <semanticDescriptor> resources on a remote CSE
+			#	by the relatedSemantics attribute or by triples with annotation property m2m:resourceDescriptorLink in descriptor attribute.
+			#	This process shall consider the recursive links.
+			#	- If yes, the Hosting CSE shall generate an Update request primitive with itself as the Originator and with the 
+			# 		Content parameter set to the addressed <semanticDescriptor> resource representation, and send it to the <semanticValidation>
+			# 		virtual resource URI on the CSE which hosts the referenced ontology (following the ontologyRef attribute) of the addressed
+			# 		<semanticDescriptor> resource (see details in clause 7.4.48.2.3). After receiving the response primitive, i.e. 
+			# 		the validation result, go to step k. If no response primitive was received due to time-out or other exceptional cases,
+			# 		the hosting CSE shall generate a Response Status Code indicating a "TARGET_NOT_REACHABLE" error.
+			#	-  If no, perform the following steps.
+			# d) Access the semantic triples from the descriptor attribute of the received <semanticDescriptor> resource.
+			# e) Access the ontology referenced in the ontologyRef attribute of the received <semanticDescriptor> resource.
+			# 	- If the ontology referenced by the ontologyRef attribute is an external ontology, not locally hosted by the Hosting CSE,
+			# 		the Hosting CSE shall retrieve it using the corresponding protocol and identifier information specified in the ontologyRef attribute.
+			#	- If the referenced ontology cannot be retrieved within a reasonable time (as defined by a local policy), the Hosting CSE shall generate
+			# 		 a Response Status Code indicating an "ONTOLOGY_NOT_AVAILABLE" error.
+			# f) Retrieve any local linked <semanticDescriptor> resources of the received <semanticDescriptor> resource following the URI(s) in
+			#  the relatedSemantics attribute (if it exists) and the URI(s) in the triples with annotation property m2m:resourceDescriptorLink (if there are any).
+			#	- Repeat this step recursively to Retrieve any further local linked <semanticDescriptor> resources.
+			# 	- If the local linked <semanticDescriptor> resources cannot be retrieved within a reasonable time (which is subject to a local policy),
+			# 		the Hosting CSE shall generate a Response Status Code indicating a "LINKED_SEMANTICS_NOT_AVAILABLE" error.
+			# g) Retrieve the semantic triples from the descriptor attribute of the local linked <semanticDescriptor> resource.
+			# h) Retrieve the referenced ontologies of the local linked <semanticDescriptor> resources following the URI(s) in ontologyRef attribute of
+			# 	the linked <semanticDescriptor> resources; If the referenced ontologies cannot be retrieved within a reasonable time (as defined by 
+			# 	a local policy), the Hosting CSE shall generate a Response Status Code indicating an "ONTOLOGY_NOT_AVAILABLE" error.
+			# i) Combine all the semantic triples of the addressed and local linked <semanticDescriptor> resources as the set of semantic triples to be
+			# 	validated, and combine all the referenced ontologies as the set of ontologies to validate the semantic triples against.
+			# j) Check all the aspects of semantic validation according to clause 7.10.3 in oneM2M TS-0034 [50] based upon the semantic triples and 
+			# 	referenced ontology. If any problem occurs, the Hosting CSE shall generate a Response Status Code indicating an "INVALID_SEMANTICS" error.
+		
 		return Result.successResult()