+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Console] Added command to add a separator line to the screen and log (at the current log level).
 - [SCRIPTS] Added *logDivider* command.
 - [CSE] Added support for &lt;crossResourceSubscription> resource type.
+- [CSE] Added support for &lt;semanticDescriptor> resource type. Only the basic resource handling is implemented yet, no semantic functionality, validation etc.
 - [CSE] Added support for notification statistics in (*nse* and *nsi* attributes).
 - [CSE] Added support for *expirationCounter* (&lt;subscription>, &lt;crossResourceSubscription>)
 - [CONSOLE] Added continuous inspection of a resource.